requirements.txt
flask==3.0.3
flask-cors==4.0.1
anthropic>=0.40.0
edge-tts>=6.1.9
requests==2.32.3
python-dotenv==1.0.1